# src/PaperDotNet.Host/openapi.json
{
  "openapi": "3.2.0",
  "info": {
    "title": "paperdotnet | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5098"
    }
  ],
  "paths": {
...
    "/v1.0/workspaces/{workspaceId}/lists/{listId}/items/{itemId}": {
      "get": {
        "tags": [
          "Items"
        ],
        "operationId": "GetItem",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Items"
        ],
        "description": "fields are merged (null removes a value); parentId moves the item (null: the list root); contentTypeId changes its content type. Requires If-Match.",
        "operationId": "UpdateItem",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonElement"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "operationId": "DeleteItem",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/v1.0/workspaces/{workspaceId}/lists/{listId}/items/{itemId}/noteLinks": {
      "get": {
        "tags": [
          "Notes"
        ],
        "description": "The note's wiki links in order, resolved to notes where possible.",
        "operationId": "GetNoteLinks",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NoteLinksResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1.0/workspaces/{workspaceId}/lists/{listId}/items/{itemId}/backlinks": {
      "get": {
        "tags": [
          "Notes"
        ],
        "description": "Notes that link to this note, by title.",
        "operationId": "GetNoteBacklinks",
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "listId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacklinksResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BacklinksResponse": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LinkedNote"
            }
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "ItemResponse": {
        "required": [
          "id",
          "listId",
          "contentTypeId",
          "parentId",
          "isFolder",
          "createdAt",
          "createdBy",
          "updatedAt",
          "updatedBy",
          "fields",
          "@odata.etag"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "listId": {
            "type": "string",
            "format": "uuid"
          },
          "contentTypeId": {
            "type": "string",
            "format": "uuid"
          },
          "parentId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "isFolder": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdBy": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedBy": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "fields": {
            "$ref": "#/components/schemas/JsonObject"
          },
          "@odata.etag": {
            "type": "string"
          }
        },
        "description": "A list item as returned by the API. `fields` contains `title` and all field values."
      },
...
      "JsonElement": {},
...
      "JsonObject": {
        "type": "object"
      },
      "LinkedNote": {
        "required": [
          "workspaceId",
          "listId",
          "itemId",
          "title"
        ],
        "type": "object",
        "properties": {
          "workspaceId": {
            "type": "string",
            "format": "uuid"
          },
          "listId": {
            "type": "string",
            "format": "uuid"
          },
          "itemId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          }
        },
        "description": "A note the caller can read."
      },
...
      "NoteLinkResponse": {
        "required": [
          "target",
          "heading",
          "alias",
          "embed",
          "note"
        ],
        "type": "object",
        "properties": {
          "target": {
            "type": "string"
          },
          "heading": {
            "type": [
              "null",
              "string"
            ]
          },
          "alias": {
            "type": [
              "null",
              "string"
            ]
          },
          "embed": {
            "type": "boolean"
          },
          "note": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/LinkedNote"
              }
            ]
          }
        },
        "description": "A wiki link of a note; LinkedNote? NoteLinkResponse.Note is the note it points to (left out when none exists or the caller cannot read it)."
      },
      "NoteLinksResponse": {
        "required": [
          "value"
        ],
        "type": "object",
        "properties": {
          "value": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoteLinkResponse"
            }
          }
        }
      },
...
    }
  },
...
}